Rename home revision page to home
